refactor length trust
