--- ppr-ui/src/components/registration/RegistrationLengthTrust.vue
@@ -1,784 +1,0 @@
-<template>
-  <v-container fluid no-gutters class="white pa-0" v-if="summaryView">
-    <v-card flat id="length-trust-summary">
-      <h2 class="pt-2 pb-5 renewal-title" v-if="isRenewal">
-          Renewal Length and Terms
-       </h2>
-      <v-row no-gutters class="summary-header pa-2" v-else>
-        <v-col cols="auto" class="pa-2">
-          <v-icon color="darkBlue">mdi-calendar-clock</v-icon>
-          <label
-            class="pl-3"
-            v-if="registrationType === APIRegistrationTypes.SECURITY_AGREEMENT"
-          >
-            <strong>{{ regTitle }} Length and Trust Indenture</strong>
-          </label>
-          <label
-            class="pl-3"
-            v-else-if="registrationType === APIRegistrationTypes.REPAIRERS_LIEN"
-          >
-            <strong>Amount and Date of Surrender</strong>
-          </label>
-          <label class="pl-3" v-else>
-            <strong>{{ regTitle }} Length</strong>
-          </label>
-        </v-col>
-      </v-row>
-      <v-container
-        :class="{ 'invalid-message': showErrorSummary }"
-        style="padding: 40px 30px;"
-      >
-        <v-row no-gutters v-if="showErrorSummary" class="pa-6">
-          <v-col cols="auto">
-            <span :class="{ 'invalid-message': showErrorSummary }">
-              <v-icon color="error">mdi-information-outline</v-icon>
-              This step is unfinished.
-            </span>
-            <span
-              id="router-link-length-trust"
-              class="invalid-link"
-              @click="goToLengthTrust()"
-            >
-              Return to this step to complete it.
-            </span>
-          </v-col>
-        </v-row>
-        <v-row no-gutters>
-          <v-col cols="3" class="generic-label"> {{ regTitle }} Length </v-col>
-          <v-col class="summary-text">
-            {{ lengthSummary }}
-          </v-col>
-        </v-row>
-        <v-row class="pt-4" v-if="renewalView && registrationType !== APIRegistrationTypes.REPAIRERS_LIEN">
-          <v-col cols="3"></v-col>
-          <v-col cols="9" class="pl-2"><v-divider class="ml-0" /></v-col>
-        </v-row>
-        <v-row no-gutters class="pt-6" v-if="renewalView">
-          <v-col cols="3" class="generic-label">New Expiry</v-col>
-          <v-col cols="9" id="new-expiry">{{ computedExpiryDateFormatted }}</v-col>
-        </v-row>
-        <v-row v-if="renewalView && showTrustIndenture">
-          <v-col cols="3"></v-col>
-          <v-col cols="9" class="pl-2"><v-divider class="ml-0" /></v-col>
-        </v-row>
-        <v-row no-gutters class="pt-6" v-if="showTrustIndenture">
-          <v-col cols="3" class="generic-label">
-            Trust Indenture
-          </v-col>
-          <v-col class="summary-text">
-            {{ trustIndentureSummary }}
-          </v-col>
-        </v-row>
-        <v-row
-          no-gutters
-          class="pt-6"
-          v-if="registrationType === APIRegistrationTypes.REPAIRERS_LIEN"
-        >
-          <v-col cols="3" class="generic-label">
-            Amount of Lien
-          </v-col>
-          <v-col class="summary-text">
-            {{ lienAmountSummary }}
-          </v-col>
-        </v-row>
-        <v-row
-          no-gutters
-          class="pt-6"
-          v-if="registrationType === APIRegistrationTypes.REPAIRERS_LIEN"
-        >
-          <v-col cols="3" class="generic-label">
-            Surrender Date
-          </v-col>
-          <v-col class="summary-text">
-            {{ surrenderDateSummary }}
-          </v-col>
-        </v-row>
-      </v-container>
-    </v-card>
-  </v-container>
-
-  <v-container fluid no-gutters class="white pa-0"
-  v-else-if="renewalView && registrationType === APIRegistrationTypes.REPAIRERS_LIEN">
-    <v-card flat id="length-trust-summary">
-      <v-row no-gutters class="summary-header pa-2">
-        <v-col cols="auto" class="pa-2">
-          <v-icon color="darkBlue">mdi-calendar-clock</v-icon>
-          <label class="pl-3">
-            <strong>Renewal Length and Terms</strong>
-          </label>
-        </v-col>
-      </v-row>
-      <v-container style="padding: 40px 30px;">
-        <v-row no-gutters>
-          <v-col cols="12" class="pb-8">
-            The length of a Repairers Lien renewal is automatically set to 180 days.
-            The registration renewal length will
-            be added to any time remaining on your current registration.
-          </v-col>
-        </v-row>
-        <v-row no-gutters>
-          <v-col cols="3" class="generic-label"> {{ regTitle }} Length </v-col>
-          <v-col class="summary-text">
-            {{ lengthSummary }}
-          </v-col>
-        </v-row>
-        <v-row no-gutters class="py-6">
-           <v-col cols="3"></v-col>
-           <v-col cols="9" class="pl-2"><v-divider class="ml-0" /></v-col>
-        </v-row>
-        <v-row no-gutters>
-          <v-col cols="3" class="generic-label"> New Expiry </v-col>
-          <v-col class="summary-text" id="new-expiry-rl">
-            {{ computedExpiryDateFormatted }}
-          </v-col>
-        </v-row>
-        <v-row no-gutters class="pt-6">
-           <v-col cols="3"></v-col>
-           <v-col cols="9" class="pl-2"><v-divider class="ml-0" /></v-col>
-        </v-row>
-        <v-row no-gutters class="pt-6">
-          <v-col cols="3" class="generic-label">
-            Amount of Lien
-          </v-col>
-          <v-col class="summary-text">
-            {{ lienAmountSummary }}
-          </v-col>
-        </v-row>
-        <v-row no-gutters class="pt-6">
-          <v-col cols="3" class="generic-label">
-            Surrender Date
-          </v-col>
-          <v-col class="summary-text" id="surrender-date">
-            {{ surrenderDateSummary }}
-          </v-col>
-        </v-row>
-      </v-container>
-    </v-card>
-  </v-container>
-  <v-container
-    fluid
-    no-gutters
-    class="white pb-6 pr-10 pl-8 rounded"
-    :class="{ 'invalid-message': lengthTrust.showInvalid }"
-    v-else
-  >
-  <v-row no-gutters v-if="renewalView" class="summary-header pa-2 mb-8 mt-n3 mr-n10 ml-n8">
-        <v-col cols="auto" class="pa-2">
-          <v-icon color="darkBlue">mdi-calendar-clock</v-icon>
-          <label class="pl-3">
-            <strong>Renewal Length <span v-if="showTrustIndenture">and Trust Indenture</span>
-            <span v-if="registrationType === APIRegistrationTypes.REPAIRERS_LIEN">and Terms</span></strong>
-          </label>
-        </v-col>
-      </v-row>
-
-        <v-row v-if="renewalView && registrationType === APIRegistrationTypes.REPAIRERS_LIEN" no-gutters>
-          <v-col cols="12" class="pb-8">
-            The length of a Repairers Lien is automatically set to 180 days. The registration renewal length will
-            be added to any time remaining on your current registration.
-          </v-col>
-        </v-row>
-
-      <v-row v-if="renewalView && registrationType !== APIRegistrationTypes.REPAIRERS_LIEN" no-gutters>
-          <v-col cols="12" class="pb-2">
-            The registration length entered below will be added to any time remaining on your
-            current registration.
-          </v-col>
-        </v-row>
-
-    <div v-if="registrationType === APIRegistrationTypes.REPAIRERS_LIEN">
-      <v-row no-gutters class="ps-6 pt-6 pb-3">
-        <v-col cols="3" class="generic-label"> {{ regTitle }} Length </v-col>
-        <v-col class="summary-text pl-4">
-          {{ lengthSummary }}
-        </v-col>
-      </v-row>
-      <v-row no-gutters class="ps-6 pt-6">
-        <v-col cols="3" class="generic-label pt-3">
-          <span :class="{ 'invalid-message': showErrorLienAmount }"
-            >Amount of Lien</span
-          >
-        </v-col>
-        <v-col>
-          <v-text-field
-            id="lien-amount"
-            autocomplete="off"
-            :error-messages="lienAmountMessage || ''"
-            filled
-            hint="Example: 10,500.50"
-            persistent-hint
-            label="Amount in Canadian Dollars ($)"
-            v-model="lienAmount"
-          />
-        </v-col>
-      </v-row>
-      <v-row no-gutters class="ps-6 pt-4">
-        <v-col cols="3" class="generic-label pt-3">
-          <span :class="{ 'invalid-message': showErrorSurrenderDate }"
-            >Surrender Date</span
-          >
-        </v-col>
-        <v-col>
-          <v-dialog
-            ref="dialog"
-            v-model="modal"
-            :return-value.sync="surrenderDate"
-            persistent
-            width="450px"
-          >
-            <template v-slot:activator="{ on, attrs }">
-              <v-text-field
-                id="date-text-field"
-                :value="computedDateFormatted"
-                :error-messages="surrenderDateMessage || ''"
-                filled
-                hint="Must be within the last 21 days"
-                persistent-hint
-                label="Date"
-                append-icon="mdi-calendar"
-                readonly
-                v-bind="attrs"
-                v-on="on"
-                v-on:click:append="on.click"
-              >
-              </v-text-field>
-            </template>
-            <v-date-picker
-              id="date-picker-calendar"
-              v-model="surrenderDate"
-              elevation="15"
-              :min="minSurrenderDate"
-              scrollable
-              width="450px"
-            >
-              <v-spacer></v-spacer>
-              <v-btn
-                text
-                color="primary"
-                @click="$refs.dialog.save(surrenderDate)"
-              >
-                <strong>OK</strong>
-              </v-btn>
-              <v-btn text color="primary" @click="modal = false">
-                Cancel
-              </v-btn>
-            </v-date-picker>
-          </v-dialog>
-        </v-col>
-      </v-row>
-    </div>
-    <div v-else>
-      <v-row class="pt-6" no-gutters>
-        <v-col cols="3" class="generic-label">
-          <span :class="{ 'invalid-message': lengthTrust.showInvalid }"
-            >{{ regTitle }} Length</span
-          >
-        </v-col>
-        <v-col cols="auto">
-          <span v-if="infinityPreselected()">
-            Infinite
-          </span>
-          <v-radio-group v-else v-model="lifeInfinite">
-            <v-radio
-              class="years-radio pa-0 ma-0"
-              :hide-details="false"
-              label=""
-              value="false"
-              id="length-in-years"
-              @click="setLifeInfinite(false)"
-            >
-            </v-radio>
-            <v-radio
-              class="infinite-radio pt-15 ma-0"
-              :hide-details="false"
-              label=""
-              value="true"
-              id="length-infinite"
-              @click="setLifeInfinite(true)"
-            >
-            </v-radio>
-          </v-radio-group>
-        </v-col>
-        <v-col v-if="!infinityPreselected()">
-          <v-text-field
-            id="life-years-field"
-            autocomplete="off"
-            :error-messages="lifeYearsMessage || ''"
-            filled
-            :readonly="lifeYearsDisabled"
-            :hint="lifeYearsHint"
-            persistent-hint
-            label="Length in Years"
-            v-model="lifeYearsEdit"
-          />
-          <div class="pt-5">Infinite ($500.00 non-refundable)</div>
-        </v-col>
-      </v-row>
-      <v-row>
-        <v-col cols="3"></v-col>
-        <v-col cols="9" class="pl-2" v-if="!infinityPreselected()"><v-divider class="ml-0" /></v-col>
-      </v-row>
-      <v-row no-gutters class="py-6" v-if="renewalView">
-        <v-col cols="3" class="generic-label">New Expiry</v-col>
-        <v-col cols="9" id="new-expiry">{{ computedExpiryDateFormatted }}</v-col>
-      </v-row>
-      <v-row v-if="renewalView && showTrustIndenture">
-        <v-col cols="3"></v-col>
-        <v-col cols="9" class="pl-2"><v-divider class="ml-0" /></v-col>
-      </v-row>
-      <v-row no-gutters class="pt-6" v-if="showTrustIndenture">
-        <v-col cols="3" class="generic-label">
-          Trust Indenture
-        </v-col>
-        <v-col class="summary-text" v-if="renewalView">
-            {{ trustIndentureSummary }}
-        </v-col>
-        <v-col cols="auto" v-if="!renewalView">
-          <v-checkbox
-            class="trust-checkbox pa-0 ma-0"
-            :hide-details="false"
-            :hint="trustIndentureHint"
-            label=""
-            id="trust-indenture-checkbox"
-            v-model="trustIndenture"
-          >
-          </v-checkbox>
-        </v-col>
-        <v-col cols="8" v-if="!renewalView">
-          <v-tooltip
-            top
-            content-class="top-tooltip pa-5"
-            transition="fade-transition"
-          >
-            <template v-slot:activator="{ on }">
-              <span v-on="on" class="trust-indenture">Trust Indenture</span>
-            </template>
-            Select if the security interest is contained in a Trust Indenture.
-          </v-tooltip>
-        </v-col>
-      </v-row>
-    </div>
-  </v-container>
-</template>
-
-<script lang="ts">
-// external
-import {
-  computed,
-  defineComponent,
-  reactive,
-  toRefs,
-  watch,
-  onMounted
-} from '@vue/composition-api'
-import { useGetters, useActions } from 'vuex-composition-helpers'
-
-// local
-import { LengthTrustIF } from '@/interfaces' // eslint-disable-line no-unused-vars
-import { convertDate } from '@/utils'
-import { APIRegistrationTypes, RouteNames } from '@/enums'
-import { getFinancingFee } from '@/composables/fees/factories'
-
-export default defineComponent({
-  props: {
-    defaultRegistrationType: {
-      type: String,
-      default: ''
-    },
-    isSummary: {
-      type: Boolean,
-      default: false
-    },
-    isRenewal: {
-      type: Boolean,
-      default: false
-    }
-  },
-  setup (props, context) {
-    const { setLengthTrust } = useActions<any>(['setLengthTrust'])
-    const { getLengthTrust } = useGetters<any>(['getLengthTrust'])
-    const { getRegistrationType, getRegistrationExpiryDate, getRegistrationSurrenderDate } = useGetters<any>([
-      'getRegistrationType', 'getRegistrationExpiryDate', 'getRegistrationSurrenderDate'
-    ])
-    const registrationType = getRegistrationType.value?.registrationTypeAPI
-    const lengthTrust: LengthTrustIF = getLengthTrust.value
-    const feeInfoYears = getFinancingFee(false)
-    const router = context.root.$router
-    const route = context.root.$route
-    const modal = false
-
-    if (
-      registrationType === APIRegistrationTypes.REPAIRERS_LIEN &&
-      lengthTrust.lifeYears !== 1 &&
-      !props.isSummary
-    ) {
-      lengthTrust.valid = true
-      lengthTrust.lifeYears = 1
-      setLengthTrust(lengthTrust)
-    }
-
-    const localState = reactive({
-      summaryView: props.isSummary,
-      renewalView: props.isRenewal,
-      trustIndenture: lengthTrust.trustIndenture,
-      lifeYearsDisabled: lengthTrust.lifeInfinite,
-      lifeInfinite: lengthTrust.valid
-        ? lengthTrust.lifeInfinite.toString()
-        : '',
-      maxYears: feeInfoYears.quantityMax.toString(),
-      lifeYearsEdit:
-        lengthTrust.lifeYears > 0 ? lengthTrust.lifeYears.toString() : '',
-      lifeYearsMessage: '',
-      trustIndentureHint: '',
-      surrenderDate: lengthTrust.surrenderDate,
-      lienAmount: lengthTrust.lienAmount,
-      lifeYearsHint:
-        'Minimum 1 year, Maximum ' +
-        feeInfoYears.quantityMax.toString() +
-        ' years ($' +
-        feeInfoYears.feeAmount.toFixed(2) +
-        ' per year)',
-      showTrustIndenture: computed((): boolean => {
-        if (localState.renewalView) {
-          return lengthTrust.trustIndenture
-        }
-        return registrationType === APIRegistrationTypes.SECURITY_AGREEMENT
-      }),
-      showErrorSummary: computed((): boolean => {
-        return !lengthTrust.valid
-      }),
-      showErrorLienAmount: computed((): boolean => {
-        if (
-          lengthTrust.lienAmount !== '' &&
-          localState.lienAmountMessage?.length > 0
-        ) {
-          return true
-        }
-        return lengthTrust.showInvalid && lengthTrust.lienAmount === ''
-      }),
-      showErrorSurrenderDate: computed((): boolean => {
-        return lengthTrust.showInvalid && lengthTrust.surrenderDate === ''
-      }),
-      lienAmountMessage: computed((): string => {
-        if (lengthTrust.showInvalid && lengthTrust.lienAmount === '') {
-          return 'This field is required'
-        }
-        if (lengthTrust.lienAmount?.length > 0 && !validLienAmount(lengthTrust.lienAmount)) {
-          return 'Lien amount must be a number greater than 0.'
-        }
-        return ''
-      }),
-      surrenderDateMessage: computed((): string => {
-        if (lengthTrust.showInvalid && lengthTrust.surrenderDate === '') {
-          return 'This field is required'
-        }
-        return ''
-      }),
-      regTitle: computed((): string => {
-        if (props.isRenewal) {
-          return 'Renewal'
-        }
-        return 'Registration'
-      }),
-      minSurrenderDate: computed((): string => {
-        var dateOffset = 24 * 60 * 60 * 1000 * 21 // 21 days in milliseconds
-        var minDate = new Date()
-        minDate.setTime(minDate.getTime() - dateOffset)
-        return minDate.toISOString()
-      }),
-      computedDateFormatted: computed((): string => {
-        return lengthTrust.surrenderDate !== ''
-          ? convertDate(new Date(lengthTrust.surrenderDate + 'T09:00:00Z'), false, false) : ''
-      }),
-      computedExpiryDateFormatted: computed((): string => {
-        if (props.isRenewal) {
-          if (lengthTrust.lifeInfinite) {
-            return 'No Expiry'
-          }
-          if ((getRegistrationExpiryDate.value) && ((registrationType === APIRegistrationTypes.REPAIRERS_LIEN))) {
-            const expiryDate = getRegistrationExpiryDate.value
-            const newExpDate = new Date(expiryDate)
-            newExpDate.setDate(newExpDate.getDate() + 180)
-            return convertDate(newExpDate, true, true)
-          }
-          if ((getRegistrationExpiryDate.value) && (parseInt(localState.lifeYearsEdit) > 0)) {
-            const expiryDate = getRegistrationExpiryDate.value
-            const numYears = parseInt(localState.lifeYearsEdit)
-            const newExpDate = new Date(expiryDate)
-            newExpDate.setFullYear(newExpDate.getFullYear() + numYears)
-            return convertDate(newExpDate, true, true)
-          }
-          return '-'
-        }
-      }),
-      lengthSummary: computed((): string => {
-        if (registrationType === APIRegistrationTypes.REPAIRERS_LIEN) {
-          return '180 Days'
-        }
-        if (!lengthTrust.lifeInfinite && lengthTrust.lifeYears < 1) {
-          return 'Not entered'
-        }
-        if (lengthTrust.lifeInfinite) {
-          return 'Infinite'
-        }
-        if (lengthTrust.lifeYears === 1) {
-          return lengthTrust.lifeYears.toString() + ' Year'
-        }
-        return lengthTrust.lifeYears.toString() + ' Years'
-      }),
-      trustIndentureSummary: computed((): string => {
-        return lengthTrust.trustIndenture ? 'Yes' : 'No'
-      }),
-      lienAmountSummary: computed((): string => {
-        if (lengthTrust.lienAmount) {
-          // Format as CDN currency.
-          var currency = lengthTrust.lienAmount
-            ?.replace('$', '')
-            ?.replaceAll(',', '')
-          var lienFloat = parseFloat(currency)
-          if (isNaN(lienFloat)) {
-            return lengthTrust.lienAmount
-          }
-          return '$' + lienFloat.toFixed(2).replace(/\d(?=(\d{3})+\.)/g, '$&,')
-        }
-        return 'Not entered'
-      }),
-      surrenderDateSummary: computed((): string => {
-        if (props.isRenewal) {
-          lengthTrust.surrenderDate = getRegistrationSurrenderDate.value
-          return convertDate(
-            new Date(lengthTrust.surrenderDate),
-            false,
-            false
-          )
-        }
-        if (lengthTrust.surrenderDate?.length >= 10) {
-          return convertDate(
-            new Date(lengthTrust.surrenderDate + 'T09:00:00Z'),
-            false,
-            false
-          )
-        }
-        if (lengthTrust.surrenderDate === '') {
-          return 'Not entered'
-        }
-        return lengthTrust.surrenderDate
-      })
-    })
-    const goToLengthTrust = (): void => {
-      if (!props.isRenewal) {
-        lengthTrust.showInvalid = true
-        setLengthTrust(lengthTrust)
-        router.push({ path: '/new-registration/length-trust' })
-      } else {
-        const registrationNumber = route.query['reg-num'] as string || ''
-        router.push({
-          name: RouteNames.RENEW_REGISTRATION,
-          query: { 'reg-num': registrationNumber }
-        })
-      }
-    }
-
-    const infinityPreselected = (): boolean => {
-      const ipArray = [
-        APIRegistrationTypes.MARRIAGE_MH,
-        APIRegistrationTypes.LAND_TAX_LIEN,
-        APIRegistrationTypes.MANUFACTURED_HOME_LIEN,
-        APIRegistrationTypes.INSURANCE_PREMIUM_TAX,
-        APIRegistrationTypes.PETROLEUM_NATURAL_GAS_TAX,
-        APIRegistrationTypes.FOREST,
-        APIRegistrationTypes.LOGGING_TAX,
-        APIRegistrationTypes.CARBON_TAX,
-        APIRegistrationTypes.RURAL_PROPERTY_TAX,
-        APIRegistrationTypes.PROVINCIAL_SALES_TAX,
-        APIRegistrationTypes.INCOME_TAX,
-        APIRegistrationTypes.MOTOR_FUEL_TAX,
-        APIRegistrationTypes.EXCISE_TAX,
-        APIRegistrationTypes.LIEN_UNPAID_WAGES,
-        APIRegistrationTypes.HERITAGE_CONSERVATION_NOTICE,
-        APIRegistrationTypes.PROCEEDS_CRIME_NOTICE,
-        APIRegistrationTypes.MAINTENANCE_LIEN,
-        APIRegistrationTypes.MANUFACTURED_HOME_NOTICE,
-        APIRegistrationTypes.OTHER,
-        APIRegistrationTypes.MINERAL_LAND_TAX,
-        APIRegistrationTypes.PROPERTY_TRANSFER_TAX,
-        APIRegistrationTypes.SCHOOL_ACT
-      ]
-      return ipArray.includes(registrationType)
-    }
-
-    onMounted(() => {
-      if (infinityPreselected()) {
-        lengthTrust.valid = true
-        lengthTrust.lifeInfinite = true
-        lengthTrust.lifeYears = null
-        setLengthTrust(lengthTrust)
-      }
-    })
-
-    const setLifeInfinite = (val: boolean): void => {
-      lengthTrust.lifeInfinite = val
-      localState.lifeYearsDisabled = lengthTrust.lifeInfinite
-      if (lengthTrust.lifeInfinite) {
-        localState.lifeYearsEdit = ''
-        lengthTrust.lifeYears = null
-        lengthTrust.valid = true
-        lengthTrust.showInvalid = false
-        setLengthTrust(lengthTrust)
-      } else {
-        lengthTrust.valid = false
-        lengthTrust.lifeYears = 0
-        setLengthTrust(lengthTrust)
-      }
-    }
-
-    const validLienAmount = (val: string): boolean => {
-      if (!val || val === '') {
-        return false
-      }
-      var lienAmount = val.trimRight().trimLeft()
-      if (lienAmount.length === 0) {
-        return false
-      }
-      if (lienAmount.startsWith('$')) {
-        lienAmount = lienAmount.substr(1)
-      }
-      var amount = Number(lienAmount.replace(/,/g, ''))
-      if (isNaN(amount) || amount < 0.01) {
-        return false
-      }
-      return true
-    }
-
-    watch(
-      () => localState.lifeYearsEdit,
-      (val: string) => {
-        localState.lifeYearsMessage = ''
-        if (val?.length > 0) {
-          var life = parseInt(val)
-          if (isNaN(life)) {
-            localState.lifeYearsMessage =
-              'Registration length must be a number between 1 and ' +
-              localState.maxYears
-            lengthTrust.valid = false
-          } else if (life < 1 || life > feeInfoYears.quantityMax) {
-            localState.lifeYearsMessage =
-              'Registration length must be between 1 and ' + localState.maxYears
-            lengthTrust.valid = false
-          } else {
-            lengthTrust.lifeYears = life
-            lengthTrust.valid = true
-            localState.lifeInfinite = 'false'
-            lengthTrust.showInvalid = false
-            setLengthTrust(lengthTrust)
-          }
-        } else {
-          if (!lengthTrust.lifeInfinite) {
-            lengthTrust.lifeYears = 0
-            setLengthTrust(lengthTrust)
-            lengthTrust.valid = false
-          }
-        }
-        if (!lengthTrust.valid && !lengthTrust.lifeInfinite) {
-          lengthTrust.valid = false
-          setLengthTrust(lengthTrust)
-        }
-      }
-    )
-    watch(
-      () => localState.trustIndenture,
-      (val: boolean) => {
-        lengthTrust.trustIndenture = val
-        setLengthTrust(lengthTrust)
-      }
-    )
-    watch(
-      () => localState.lienAmount,
-      (val: string) => {
-        lengthTrust.lienAmount = val.trimRight().trimLeft()
-        if (!validLienAmount(val)) {
-          lengthTrust.valid = false
-        } else if (lengthTrust.surrenderDate !== '') {
-          lengthTrust.valid = true
-          lengthTrust.showInvalid = false
-        }
-        setLengthTrust(lengthTrust)
-      }
-    )
-    watch(
-      () => localState.surrenderDate,
-      (val: string) => {
-        lengthTrust.surrenderDate = val
-        if (lengthTrust.lienAmount !== '' && lengthTrust.surrenderDate !== '') {
-          lengthTrust.valid = true
-          lengthTrust.showInvalid = false
-        } else {
-          lengthTrust.valid = false
-        }
-        setLengthTrust(lengthTrust)
-      }
-    )
-
-    return {
-      goToLengthTrust,
-      setLifeInfinite,
-      lengthTrust,
-      infinityPreselected,
-      APIRegistrationTypes,
-      registrationType,
-      modal,
-      ...toRefs(localState)
-    }
-  }
-})
-</script>
-
-<style lang="scss" scoped>
-/* Need scoped for date picker v-deep style overrides to work */
-@import '@/assets/styles/theme.scss';
-.v-list-item {
-  min-height: 0;
-}
-
-.renewal-title {
-   background-color: #f1f3f5;
-}
-
-::v-deep
-  .v-icon.v-icon:not(.mdi-radiobox-marked):not(.mdi-radiobox-blank):not(.mdi-checkbox-blank-outline) {
-  color: $primary-blue;
-}
-::v-deep .v-picker__title__btn:not(.v-picker__title__btn--active) {
-  opacity: 1;
-}
-::v-deep .v-date-picker-table__current {
-  border-color: $primary-blue !important;
-}
-::v-deep .v-date-picker-table__current .v-btn__content {
-  color: $primary-blue !important;
-}
-::v-deep .theme--light.v-date-picker-table th {
-  color: $gray9;
-}
-::v-deep .v-date-picker-table .v-btn {
-  color: $gray7;
-}
-::v-deep
-  .theme--light.v-btn:not(.v-btn--flat):not(.v-btn--text):not(.v-btn--outlined) {
-  background-color: $primary-blue !important;
-  border-color: $primary-blue !important;
-  color: white !important;
-}
-::v-deep .v-btn:not(.v-btn--text):not(.v-btn--outlined).v-btn--active:before {
-  opacity: 0;
-}
-::v-deep .v-icon.v-icon.v-icon--link {
-  cursor: text;
-}
-::v-deep .theme--light.v-icon.v-icon.v-icon--disabled {
-  color: $primary-blue !important;
-}
-::v-deep .v-input--is-disabled {
-  opacity: 0.4;
-}
-</style>
